requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
boto3==1.34.107
botocore==1.34.107
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
s3transfer==0.10.1
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
